+++ apim-billing-engine/src/main/java/org/wso2/apim/billing/services/impl/MeteredPlanProcessor.java
@@ -27,11 +27,11 @@ public class MeteredPlanProcessor extends PlanProcessor {
             } else {
                 return null;
             }
-            long totalRequestCount = usage.getSuccessCount() + usage.getExceedCount();
-            long pricePerReq = getValue(billingModel.getAttributes(), "pricePerReq");
-            long subscriptionFee = getValue(billingModel.getAttributes(), "subscription");
-            long usageFee = totalRequestCount * pricePerReq;
-            long totalCost = subscriptionFee + usageFee;
+            double totalRequestCount = usage.getSuccessCount() + usage.getExceedCount();
+            double pricePerReq = getValue(billingModel.getAttributes(), "pricePerReq");
+            double subscriptionFee = getValue(billingModel.getAttributes(), "subscription");
+            double usageFee = totalRequestCount * pricePerReq;
+            double totalCost = subscriptionFee + usageFee;
             model.setTotalCost(totalCost);
             List<BillingAttribute> attributes = new ArrayList<>(billingModel.getAttributes());
             attributes.add(new BillingAttribute(0, "usageFee", "Usage Fee", String.valueOf(usageFee)));
